# Improve the performance of string gathers under ugni