This partially reverts #275. In #275 the string gather implementation was
split to optimize for ugni vs. non-ugni configurations. At the time pure
unorderedCopy was faster than localizing and using aggregation but
that's no longer true.

Back then, `newSrcAggregator` actually mapped to `unorderedCopy` under
ugni because of scalabiltiy isues with the SrcAggregator under ugni. In
#507 that was changed so that `newSrcAggregator` actually maps to
`SrcAggregator` now. That provided significant speedups for gathering
remote things, especially single-byte items like the `uint(8)` used to
implemented segmented strings. There have also been performance
improvements to the DstAggregator since then. All that to say that while
in the #275 timeframe unorderedCopy was faster for string gathers, that
is no longer the case, so use aggregation.

Here's some performance results from the machine we run the nightly
16-node-xc testing on (the old implementation took too long for the
large problem size):

```
make test-bin/StringGatherSpeedTest
./test-bin/StringGatherSpeedTest --N=1_000_000   --MEANLEN=100 -nl 16
./test-bin/StringGatherSpeedTest --N=100_000_000 --MEANLEN=100 -nl 16
```

| --N         | before    | now         |
| ----------: | --------: | ----------: |
|   1_000_000 | 25.5 MB/s |  827.5 MB/s |
| 100_000_000 | N/A       | 1855.5 MB/s |


Part of https://github.com/mhmerrill/arkouda/issues/527